Forward reference error includes line numbers
