--- tests/untried/neg/forward.scala
@@ -1,24 +1,0 @@
-object Test {
-  def f: Int = x;
-  val x: Int = f;
-
-  {
-    def f: Int = x;
-    val x: Int = f;
-  }
-  {
-    def f: Int = g;
-    val x: Int = f;
-    def g: Int = x;
-  }
-  {
-    def f: Int = g;
-    var x: Int = f;
-    def g: Int = x;
-  }
-  {
-    def f: Int = g;
-    Console.println("foo");
-    def g: Int = f;
-  }
-}
